+++ front/cypress/e2e/idea_posting_anonymous.cy.ts
@@ -1,61 +1,218 @@
-import { randomString, randomEmail } from '../support/commands';
+import moment = require('moment');
+import { randomString } from '../support/commands';
 
-describe('Idea posting anonymous', () => {
-  let projectId = '';
-  let projectSlug = '';
-
+describe('Continuous ideation with anonymous participation allowed', () => {
   const projectTitle = randomString();
-  const projectDescriptionPreview = randomString();
-  const projectDescription = randomString();
-
-  const firstName = randomString();
-  const lastName = randomString();
-  const email = randomEmail();
-  const password = randomString();
+  const projectDescriptionPreview = randomString(30);
+  let projectId: string;
+  let projectSlug: string;
 
   before(() => {
-    cy.apiCreateProject({
-      type: 'continuous',
-      title: projectTitle,
-      descriptionPreview: projectDescriptionPreview,
-      description: projectDescription,
-      publicationStatus: 'published',
-      participationMethod: 'ideation',
-      allow_anonymous_participation: true,
-    }).then((response) => {
-      projectId = response.body.data.id;
-      projectSlug = response.body.data.attributes.slug;
-    });
-
-    cy.apiSignup(firstName, lastName, email, password);
+    cy.setAdminLoginCookie();
+    cy.getAuthUser()
+      .then((user) => {
+        return cy.apiCreateProject({
+          type: 'continuous',
+          title: projectTitle,
+          descriptionPreview: projectDescriptionPreview,
+          description: randomString(),
+          publicationStatus: 'published',
+          assigneeId: user.body.data.id,
+          participationMethod: 'ideation',
+          maxBudget: 400,
+          allow_anonymous_participation: true,
+        });
+      })
+      .then((project) => {
+        projectId = project.body.data.id;
+        projectSlug = project.body.data.attributes.slug;
+      });
   });
 
-  beforeEach(() => {
-    cy.setLoginCookie(email, password);
+  it('admin can submit anonymous idea', () => {
+    const ideaTitle = randomString(20);
+    const ideaContent = randomString(60);
+
+    cy.setAdminLoginCookie();
+
     cy.visit(`/projects/${projectSlug}/ideas/new`);
+
+    cy.get('#e2e-idea-new-page');
     cy.get('#idea-form');
-    cy.acceptCookies();
-  });
+    cy.contains('Add new idea').should('exist');
 
-  it('lets you post anonymously', () => {
-    const ideaTitle = randomString();
-    const ideaContent = randomString(300);
-
+    // add a title and description
     cy.get('#e2e-idea-title-input input').type(ideaTitle);
     cy.get('#e2e-idea-description-input .ql-editor').type(ideaContent);
 
-    cy.get('.e2e-checkbox').first().click();
+    // set to anonymous
+    cy.get('#e2e-post-anonymously-checkbox').should('exist');
+    cy.get('#e2e-post-anonymously-checkbox').click();
+
+    // save the form
     cy.get('.e2e-submit-idea-form').click();
+
+    // confirm anonymous participation
     cy.get('#e2e-continue-anonymous-participation-btn').should('exist');
     cy.get('#e2e-continue-anonymous-participation-btn').click();
 
-    cy.get('#e2e-idea-show');
+    // verify the content of the newly created idea page
     cy.location('pathname').should('eq', `/en/ideas/${ideaTitle}`);
+    cy.get('#e2e-idea-show');
+    cy.get('#e2e-idea-show').find('#e2e-idea-title').contains(ideaTitle);
 
-    cy.get('.e2e-modal-close-button').should('exist');
-    cy.get('.e2e-modal-close-button').first().click();
+    // verify that the author is anonymous
+    cy.get('#e2e-anonymous-username').should('exist');
+  });
 
-    cy.get('.e2e-username').should('include.text', 'Anonymous');
+  it('resident can submit anonymous idea', () => {
+    const ideaTitle = randomString(20);
+    const ideaContent = randomString(60);
+
+    cy.setLoginCookie('user@citizenlab.co', 'democracy2.0');
+
+    cy.visit(`/projects/${projectSlug}/ideas/new`);
+
+    cy.get('#e2e-idea-new-page');
+    cy.get('#idea-form');
+    cy.contains('Add new idea').should('exist');
+
+    // add a title and description
+    cy.get('#e2e-idea-title-input input').type(ideaTitle);
+    cy.get('#e2e-idea-description-input .ql-editor').type(ideaContent);
+
+    // set to anonymous
+    cy.get('#e2e-post-anonymously-checkbox').should('exist');
+    cy.get('#e2e-post-anonymously-checkbox').click();
+
+    // save the form
+    cy.get('.e2e-submit-idea-form').click();
+
+    // confirm anonymous participation
+    cy.get('#e2e-continue-anonymous-participation-btn').should('exist');
+    cy.get('#e2e-continue-anonymous-participation-btn').click();
+
+    // verify the content of the newly created idea page
+    cy.location('pathname').should('eq', `/en/ideas/${ideaTitle}`);
+    cy.get('#e2e-idea-show');
+    cy.get('#e2e-idea-show').find('#e2e-idea-title').contains(ideaTitle);
+
+    // verify that the author is anonymous
+    cy.get('#e2e-anonymous-username').should('exist');
+  });
+
+  after(() => {
+    cy.apiRemoveProject(projectId);
+  });
+});
+
+describe('Timeline ideation with anonymous participation allowed', () => {
+  const projectTitle = randomString();
+  const projectDescriptionPreview = randomString(30);
+  let projectId: string;
+  let projectSlug: string;
+
+  before(() => {
+    cy.setAdminLoginCookie();
+    cy.getAuthUser().then(() => {
+      cy.apiCreateProject({
+        type: 'timeline',
+        title: projectTitle,
+        descriptionPreview: '',
+        description: '',
+        publicationStatus: 'published',
+      }).then((project) => {
+        projectId = project.body.data.id;
+        projectSlug = project.body.data.attributes.slug;
+        cy.apiCreatePhase(
+          projectId,
+          'phaseTitle',
+          moment().subtract(2, 'month').format('DD/MM/YYYY'),
+          moment().add(2, 'days').format('DD/MM/YYYY'),
+          'ideation',
+          true,
+          true,
+          true,
+          undefined,
+          undefined,
+          undefined,
+          undefined,
+          true
+        );
+      });
+    });
+  });
+
+  it('admin can submit anonymous idea', () => {
+    const ideaTitle = randomString(20);
+    const ideaContent = randomString(60);
+
+    cy.setAdminLoginCookie();
+
+    cy.visit(`/projects/${projectSlug}/ideas/new`);
+
+    cy.get('#e2e-idea-new-page');
+    cy.get('#idea-form');
+    cy.contains('Add new idea').should('exist');
+
+    // add a title and description
+    cy.get('#e2e-idea-title-input input').type(ideaTitle);
+    cy.get('#e2e-idea-description-input .ql-editor').type(ideaContent);
+
+    // set to anonymous
+    cy.get('#e2e-post-anonymously-checkbox').should('exist');
+    cy.get('#e2e-post-anonymously-checkbox').click();
+
+    // save the form
+    cy.get('.e2e-submit-idea-form').click();
+
+    // confirm anonymous participation
+    cy.get('#e2e-continue-anonymous-participation-btn').should('exist');
+    cy.get('#e2e-continue-anonymous-participation-btn').click();
+
+    // verify the content of the newly created idea page
+    cy.location('pathname').should('eq', `/en/ideas/${ideaTitle}`);
+    cy.get('#e2e-idea-show');
+    cy.get('#e2e-idea-show').find('#e2e-idea-title').contains(ideaTitle);
+
+    // verify that the author is anonymous
+    cy.get('#e2e-anonymous-username').should('exist');
+  });
+
+  it('resident can submit anonymous idea', () => {
+    const ideaTitle = randomString(20);
+    const ideaContent = randomString(60);
+
+    cy.setLoginCookie('user@citizenlab.co', 'democracy2.0');
+
+    cy.visit(`/projects/${projectSlug}/ideas/new`);
+
+    cy.get('#e2e-idea-new-page');
+    cy.get('#idea-form');
+    cy.contains('Add new idea').should('exist');
+
+    // add a title and description
+    cy.get('#e2e-idea-title-input input').type(ideaTitle);
+    cy.get('#e2e-idea-description-input .ql-editor').type(ideaContent);
+
+    // set to anonymous
+    cy.get('#e2e-post-anonymously-checkbox').should('exist');
+    cy.get('#e2e-post-anonymously-checkbox').click();
+
+    // save the form
+    cy.get('.e2e-submit-idea-form').click();
+
+    // confirm anonymous participation
+    cy.get('#e2e-continue-anonymous-participation-btn').should('exist');
+    cy.get('#e2e-continue-anonymous-participation-btn').click();
+
+    // verify the content of the newly created idea page
+    cy.location('pathname').should('eq', `/en/ideas/${ideaTitle}`);
+    cy.get('#e2e-idea-show');
+    cy.get('#e2e-idea-show').find('#e2e-idea-title').contains(ideaTitle);
+
+    // verify that the author is anonymous
+    cy.get('#e2e-anonymous-username').should('exist');
   });
 
   after(() => {
