+++ front/cypress/e2e/proposal_posting_anonymous.cy.ts
